# backend/app/http/controllers/admin_controller.go
package controllers

import (
	"onlineJudge/backend/app/models"
	"onlineJudge/backend/database"

	"github.com/gofiber/fiber/v2"
)

// GetPendingProblems godoc
// @Summary Get pending problems
// @Description Get list of problems waiting for review
// @Tags Admin
// @Produce json
// @Success 200 {array} models.Problem
// @Router /admin/problems [get]
func GetPendingProblems(c *fiber.Ctx) error {
	role := c.Locals("role").(string)
	if role != "admin" {
		return c.Status(403).JSON(fiber.Map{"error": "Access denied"})
	}

	var problems []models.Problem
	database.DB.Where("status = ?", "pending_review").Find(&problems)
	return c.JSON(problems)
}

// ApproveProblem godoc
// @Summary Approve a problem
// @Description Approve a problem and make it public
// @Tags Admin
// @Param id path int true "Problem ID"
// @Success 200 {object} map[string]string
// @Router /admin/problems/{id}/approve [post]
func ApproveProblem(c *fiber.Ctx) error {
	role := c.Locals("role").(string)
	if role != "admin" {
		return c.Status(403).JSON(fiber.Map{"error": "Access denied"})
	}

	id := c.Params("id")
	var problem models.Problem
	if err := database.DB.First(&problem, id).Error; err != nil {
		return c.Status(404).JSON(fiber.Map{"error": "Problem not found"})
	}

	problem.Status = "approved"
	problem.Visibility = "public"
	database.DB.Save(&problem)

	return c.JSON(fiber.Map{"message": "Problem approved"})
}

// RejectProblem godoc
// @Summary Reject a problem
// @Description Reject a problem and set it back to draft
// @Tags Admin
// @Param id path int true "Problem ID"
// @Success 200 {object} map[string]string
// @Router /admin/problems/{id}/reject [post]
func RejectProblem(c *fiber.Ctx) error {
	role := c.Locals("role").(string)
	if role != "admin" {
		return c.Status(403).JSON(fiber.Map{"error": "Access denied"})
	}

	id := c.Params("id")
	var problem models.Problem
	if err := database.DB.First(&problem, id).Error; err != nil {
		return c.Status(404).JSON(fiber.Map{"error": "Problem not found"})
	}

	problem.Status = "draft"
	problem.Visibility = "private"
	database.DB.Save(&problem)

	return c.JSON(fiber.Map{"message": "Problem rejected"})
}